backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
reportlab
matplotlib
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic[email]==2.5.0
python-dotenv==1.0.0
celery==5.4.0
redis==5.0.7
pywebpush==2.1.0
